chore(compile): fix compile errors after EDC refactor
