+++ build.gradle.kts
@@ -59,10 +59,4 @@ allprojects {
         configDirectory.set(rootProject.file("resources"))
     }
 
-    // EdcRuntimeExtension uses this to determine the runtime classpath of the module to run.
-    tasks.register("printClasspath") {
-        doLast {
-            println(sourceSets["main"].runtimeClasspath.asPath)
-        }
-    }
 }
